requirements.txt
arrow==1.3.0
astroid==3.3.9
beautifulsoup4==4.13.3
blinker==1.9.0
bs4==0.0.2
certifi==2025.1.31
charset-normalizer==3.4.1
click==8.1.8
dill==0.3.9
exceptiongroup==1.2.2
Flask==3.1.0
idna==3.10
iniconfig==2.1.0
isort==6.0.1
itsdangerous==2.2.0
Jinja2==3.1.6
MarkupSafe==3.0.2
mccabe==0.7.0
mysql-connector-python==9.2.0
packaging==24.2
platformdirs==4.3.7
pluggy==1.5.0
pycodestyle==2.13.0
pydocstyle==6.3.0
pylint==3.3.6
pytest==8.3.5
python-dateutil==2.9.0.post0
requests==2.32.3
six==1.17.0
snowballstemmer==2.2.0
soupsieve==2.6
tomli==2.2.1
tomlkit==0.13.2
types-python-dateutil==2.9.0.20241206
typing_extensions==4.13.1
urllib3==2.3.0
Werkzeug==3.1.3
